function to save wether a user registered or unsubscribed the newsletter

+++ unsubscribe/index.php
@@ -1,21 +1,24 @@
 <?php
 require_once("../include/db.php");
+require_once("../include/validateEmail.php");
 
 $id = $_GET['id'];
 
-if($id != ""){
+$valid = validateEmail::checkForId($id);
+
+if($id != "" && $valid){
     $query = "delete from newsletter where id = '" . $id . "'";
     $resultNewsletter = db::getInstance()->dbquery($query);
     $query = "delete from topics where uid = '" . $id . "'";
     $resultTopics = db::getInstance()->dbquery($query);
-    $query = "INSERT INTO events (eventtype) values('unregister')";
+    $query = "INSERT INTO events (eventtype) values('unsubscribe')";
     $resultStatics = db::getInstance()->dbquery($query);
 
-    if($resultNewsletter && $resultTopics && $resultStatics == true){
+    if($resultNewsletter && $resultTopics && $resultStatics){
         echo "Sie haben sich erfolgreich vom Newsletter abgemeldet.";
     }else{
-        echo "Fehler. Die Emailadresse wurde entweder nicht gefunden, oder konnte nicht aus unserer Datenbank gelöscht werden. Bitte versuchen Sie es erneut.";
+        echo "Fehler. Die Emailadresse wurde entweder nicht gefunden, konnte nicht aus unserer Datenbank gelöscht oder konnte statistisch nicht erfasst werden. Bitte versuchen Sie es erneut.";
     }
 }else{
-    echo "Fehler. Die Emailadresse konnte nicht gelöscht werden: ID = NULL";
+    echo "Fehler. Die Emailadresse konnte nicht gelöscht werden: ID = NON_EXISTENT";
 }
